requirements.txt
# LangChain Framework
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.13
langchain-chroma>=0.1.0
langchain-groq>=0.1.0

# Vector Database
chromadb>=0.4.0

# PDF Processing
pymupdf>=1.23.0
pypdf>=3.0.0

# Data Processing
pandas>=2.0.0

# AI Model Providers  
groq>=0.4.0

# Text-to-Speech
pyttsx3>=2.90

# Web Framework
flask>=2.3.0
flask-cors>=4.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
